Add checkExpirationDate method to Token class
A new private method `checkExpirationDate` has been added to the `Token` class within the `GtvApiHub.WebApi` namespace. This method retrieves the current token settings using `_tokenSettingsManager.GetTokenSettings()`, checks if the `ExpiresIn` property is not null, and compares the expiration date to the current date and time minus one minute. It returns `true` if the token is still valid and `false` if the token is expired or invalid.