Add Price class to handle price retrieval in GtvApi service
Introduced a new class `Price` in the `GtvApiHub.WebApi.Services` namespace. This class implements the `IPrice` interface and provides methods to interact with a price collection in the GtvApi service. It includes methods to retrieve all prices (`GetAsync`) and to retrieve prices by item code (`GetByItemCodeAsync`). The class uses dependency injection to obtain an instance of `IPriceEndPoint` through the constructor, which is then used to make asynchronous calls to the service endpoints. The responses from these calls are processed to extract lists of `PriceDto` objects.