Rename GetByLanguageAsync to GetAsync in Item.cs
The method GetByLanguageAsync was renamed to GetAsync in the
Item.cs file within the GtvApiHub.WebApi.Services namespace.
This change affects the method signature, and any references
to this method will need to be updated accordingly. The
functionality of the method remains the same, as it still
retrieves a list of ItemDto objects based on the provided
LanguageCode.