Add AlternativeItemDto class for Firestore integration
Introduced a new class `AlternativeItemDto` in the `GtvApiHub.WebApi.DTOs` namespace. This class is decorated with the `[FirestoreData]` attribute, indicating it maps to a Firestore document. It implements `IBaseDto`, `IResponseDto`, and `IFirestoreDto` interfaces. The class includes four properties (`BaseItemCode`, `ItemCode`, `Match`, and `Remarks`), each decorated with the `[FirestoreProperty]` attribute. Additionally, it has two read-only properties, `CollectionName` and `DocumentUniqueField`, providing the Firestore collection name and a unique identifier for the document, respectively.