Add GtvFirebaseItemDtoBuilder for constructing Firestore DTO
A new class `GtvFirebaseItemDtoBuilder` has been added to the
`GtvApiHub.WebApi.FirebaseManagement` namespace. This class
includes a private property `_firestoreItemDto` of type
`FirestoreItemDto`. The class provides a `Build` method that
takes several lists of DTOs (`ItemDto`, `PriceDto`, `StockDto`,
`AttributeDto`, `CategoryTreeDto`, `PackageTypeDto`) and a
string `itemCode`. The `Build` method filters these lists based
on the `itemCode` and constructs a `FirestoreItemDto` object
with the filtered data. The `Build` method returns the
constructed `FirestoreItemDto`. Additionally, a `Get` method
is provided to retrieve the `_firestoreItemDto` object.