Add Stock service class implementing IStock interface
Introduce a new Stock class in the GtvApiHub.WebApi.Services namespace.
This class implements the IStock interface and provides two asynchronous
methods, GetAsync() and GetAsync(string itemCode), to fetch stock data.
The class interacts with an IStockEndPoint service to retrieve and process
the data into a list of StockDto objects.