Add FirestoreContentsDtoBuilder class for DTO construction
A new class `FirestoreContentsDtoBuilder` has been added to the
`GtvApiHub.WebApi.FirebaseManagement` namespace. This class
includes a private property `_contentsDto` of type `ContentsDto`
and a `Build` method that constructs a list of `ItemCodePair`
objects from a list of `FirestoreItemDto` objects. The `_contentsDto`
property is initialized with this list and returned by the `Build`
method. Additionally, several namespaces have been imported to
support these changes.