Handle NotFound and unsuccessful responses in GetAsync
Added a new using directive for GtvApiHub.Exceptions in GtvCategoryTree.cs. Updated the GetAsync method to handle NotFound status by returning an empty list of CategoryTreeDto. Added a check to throw an ApiException for unsuccessful responses with detailed error messages. Similar changes were made to the GetAsync method with an itemCode parameter.